rename project status to sample status fix bug in finished date
